Hmac key type and manager.
While working on this, I noticed that sometimes the same key manager
might be registered multiple times for the same type url. This happens
quite frequently in tests. So I changed the behavior of
Registry.registerKeyManager. If there is an existing manager, it doesn't
do anything, but returns false. Otherwise it inserts the manager and
returns true.

Change-Id: I257251f1696764fa87586da7c3a48bee6ffb601a
ORIGINAL_AUTHOR=Thai Duong <thaidn@gmail.com>

GitOrigin-RevId: 89ab2ff8111eb4df606719434da61e79d1352a7e