Add a EngineFactory that gives Cipher and Mac instances, following a certain provider order.
To change the default provider order, just add a static block setting defaultPolicy in EngineFactory,
or do it in a specific place with getCustomFooProvider(String...).

This integrates in MacJCE and AesCtrJceCipher as canaries, but a subsequent CL should swap every
JCE-using code to use that new mechanism.

Change-Id: I7163313369dce4a90007e8ba6db44d0c2d093059
ORIGINAL_AUTHOR=Google Engineers <noreply@google.com>

GitOrigin-RevId: 110b457dce82e9d5d07f66d048bf9212ae76f191