Stop using reflection in Tink because reflection doesn't work with ProGuard.
Reflection is used in functions that transform a private keyset to a public
keyset. At the moment only Tinkey needs this feature, so these functions
are temporarily moved to TinkeyUtil. Android apps might also need this feature
so we're working on a solution that doesn't depend on reflection.

Change-Id: I90358713215a3ab222e3b62ba5deb898b9b450a3
ORIGINAL_AUTHOR=Thai Duong <thaidn@google.com>
GitOrigin-RevId: d5e62cf596f5ec878595c8bd4ee9ac093f116193