Adding a utility to fetch the Google signing keys.
This utility takes care of caching results in memory and proactively
refreshing the keys before we get to the expiration.

This utility is also thread safe. For best performance results, this
utility should be kept in a singleton. For convenience two singleton
instances are provided already and should suffice for most use cases.

Change-Id: Ic07941f2d5cb5672ed9df36379d08fbe5ebb1dae
ORIGINAL_AUTHOR=Walter Cacau <waltercacau@google.com>
GitOrigin-RevId: 8212d3b6cb557063ef586ee61ad0252ffcce56ad