Using released tarballs or commits rather than GitHub archives.
These archives are prepared on-the-fly by GitHub and are not
guaranteed to be the same every time (for example, the order of files
within the archive could change).

We also need a new version of rules_go because the version that
we were using also depends on GitHub archives. This explains
the changes in go/.

Also switch to googletest's provided bazel BUILD file, as
recommended by the googletest team.

Change-Id: I57f9c93a41eb68ee2946a0744506a9334b6d1ccc
ORIGINAL_AUTHOR=Thai Duong <thaidn@google.com>
GitOrigin-RevId: eab348375b10cedd354da3fc2f286abc80869df9