Moving common hybrid encryption methods to subtle.EllipticCurves.
When computing the ECDH shared secret, validating that the public key's
domain paramters are the same as the private key's. Also validating
that the ECDH shared secret is a x-coordinate of a valid point on the
private key's curve.

Adding tests and removing the useless checkPublicKey function.

Change-Id: I3c5400d6f3cced3e26a7d159d09c3087f6128dd3
ORIGINAL_AUTHOR=Thai Duong <thaidn@google.com>
GitOrigin-RevId: 407b8b44ec0281c1bcf680ba3438f53ee96323ab