Replacing StringPiece with const std::string& when it makes sense.
This makes sense if the StringPiece is converted to std::string
with StringPiece.ToString() which does copy the string, thus should
be avoided.

This also makes the migration to absl::string_view easier. Somehow
one cannot pass a string_view to methods that take const std::string&
(it's possible to do that with StringPiece so I guess it's because
string_view does not provide an implicit conversion to const std::string&,
but I'm not sure).

Change-Id: Ic905a78752c63ac60f25d73ba7898f723e41d9df
ORIGINAL_AUTHOR=Thai Duong <thaidn@google.com>
GitOrigin-RevId: fbd1fb69494ac4b9b982804c86fcae1f2902aef7