Factor out every single class from monster class DjbCipher.
Rename DjbCipher to SnuffleCipher as DjbCipher is not meaningful. Note
that ChaCha and Salsa are also known as Snuffle 2008 and Snuffle 2005
accordingly.

This CL doesn't make any functional changes. Improvement, fixing bugs or
rewriting it is deferred in the follow-up CLs. Otherwise, it's impossible
to review.

Change-Id: Iba65aa12484d5ff8728ef69b99cd23188cdc1563
ORIGINAL_AUTHOR=Quan Nguyen <quannguyen@google.com>
GitOrigin-RevId: 678bbb1f91f9c2aa7c1564090b572f5c9d10d960