Fixing Kokoro build.
The build is failing probably because Kokoro upgraded its Ubuntu image.

This image doesn't have the unlimited Java policy, thus tests with
256-bit keys failed. These tests are skipped now. I also temporarily
remove tests with 256-bit keys from AEAD cross-language tests. These
tests will be reinstalled after b/35928521 is fixed.

The Go compiler on Kokoro, which is at version go1.9 linux/amd64, refuses
to compile our Go code because it found a cycle dependency. This is somehow
not a problem with the version of Go on my workstation. Fortunately, the
cycle dependency is in a test, so I temporarily removed it. I'm working on
a better fix which moves the test to a different package.

I also upgrade Bazel to 0.9.0 and use rules-apple at
a2b620070d373e4f265194b69f65e9e5c17fbcb8, instead of master, which is known
to work well with Bazel 0.9.0.

Change-Id: I97229342065f12d0eec6ff31b54922161267943a
ORIGINAL_AUTHOR=Thai Duong <thaidn@google.com>
GitOrigin-RevId: 0e96efbd943ddc3c97a2b7d573f7e20b78781a5d