Generate Maven jars with Bazel.
We no longer need to maintain Maven as a separate build and test system.
This would save us a lot of time, as Maven is not exactly nice to deal
with.

This also allows us to simplify the poms, especially the dependencies.
For example, to use Tink for Android users no longer have to set
$ANDROID_HOME.

Change-Id: I925b37409f8ed5e32f228d8de0f0845ea38bc427
ORIGINAL_AUTHOR=Google Engineers <noreply@google.com>
GitOrigin-RevId: 49470c492c6bc1d28960c243b671208863ccdf6b