Fixing the pre-Keystore -> Keystore upgrade path.
If Keystore was disabled in the past and cleartext keyset written to storage, then Keystore is enabled, trying to decrypt the cleartext keyset immediately throws a security exception. The existing code assumes that decryption will succeed and generate "junk data", throwing an InvalidProtocolBufferException.

The fix, proposed by pkorth@, is to catch the exception that is actually thrown in this edge case, and fall back to reading the cleartext keyset as intended.

While I'm here, reformatting the code and ensuring that we don't throw unchecked exception when reading keysets from private preferences.

PiperOrigin-RevId: 191108969
GitOrigin-RevId: b3b53ef357ff326689dd92e906e931adc3ddd409