Switch AesGcmBoringSsl to EVP_AEAD.
EVP_CIPHER is the legacy OpenSSL API that's kind of a disaster, hence recent
issues around null pointers and the like. New code should be written against
EVP_AEAD which is much more sensible. Notably, this file is about half as long
now.

On top of things, this should be more efficient as it allows BoringSSL to reuse
the AES key schedule. (EVP_AEAD_CTX is immutable after initialization. It can
used multiple times or concurrently or whatever.)

PiperOrigin-RevId: 211020450
GitOrigin-RevId: 0734d56e3432c24e9e75e78accefd6ae36eb9be4