Avoid reinterpret_cast.
It is better to avoid reinterpret_cast and use static_cast instead if one can. For example, if one uses multiple (even interface) inheritance, static_cast actually changes the underlying pointer value in order to make it point to the corresponding part of the object which looks like a base (or a derived) object. This cl gets rid of some of the instances.

PiperOrigin-RevId: 213262011
GitOrigin-RevId: 5302e58b363e5daa6ecb1d0a6a11217e7ce77812