Remove the main files from (hopefully) all tests.
Our tests are linked with gunit_main (or gunit_main_no_google3), which means that main is already defined in another translation unit. Having two main functions violates ODR and hence can result in undefined behavior.

There is no need to fear losing tests. If a build system tries to run a test and there is no main function, then it will fail. Hence, i cannot see a way for this to stop testing a test.

PiperOrigin-RevId: 213593596
GitOrigin-RevId: 1fd8b2dcc18d24901d77170082587557cc5303e3