Add a "NoSecretKeysetHandle" getter function.
Please review the interface carefully, as it will be hard to change in the future.

This is similar to the java class NoSecretKeysetHandle at http://google3/third_party/tink/java/src/main/java/com/google/crypto/tink/NoSecretKeysetHandle.java

The main difference is that we only provide a single function which directly takes a Keyset and then returns a KeysetHandle.

PiperOrigin-RevId: 213769723
GitOrigin-RevId: 2c39c59c6edebee550d538bf84cdadbbae20b7e7