Make the KeysetHandle copiable.
When developing tests I had some problems because I need the handle at two points, and at the moment it is not copiable. I don't see any reason for such a restriction, so I suggest we remove it. Additionally, I remove the unique_ptr there (there's really no reason for it).

Because the keyset then does not keep a pointer anymore it will delete the unique ptr passed. Because of this, I change the interface of NoSecretKeysetHandle to have users pass a key by value. [This is not backwards compatible but since it's very new I think there won't be a problem]. I think this is the only part where there is such a public interface.

PiperOrigin-RevId: 214227631
GitOrigin-RevId: ca28c621404165d00a356b37ce485e122b8b3db7