Implement EcdsaPrivateKeyFactory based on KeyFactoryBase<>.
This requires to change the inheritance from KeyFactory to virtual. The class diagram will look as follows:

            KeyFactory
               /   \
            * /     \ *
             /       \
PrivateKeyFactory  KeyFactoryBase
             \       /
              \     /
               \   /
        EcdsaPrivateKeyFactory

In order to achieve this, and not make C++ think that the class 'KeyFactory' should be treated as separate base classes, we need to mark the inheritance marked * in the above diagram as "virtual".

Virtual inheritance is allowed in C++ (though people don't like it because shows a complicated class hierarchy). There are also plenty of examples: http://cs/?q=%22+public+virtual%22

PiperOrigin-RevId: 214419799
GitOrigin-RevId: a071bc720a06cadddc23302d29fb7b59b76bd2c9