Remove the recursive mutex from the registry.
This is fairly straightforward -- the only change is that the functions get_new_key_allowed and get_key_factory are internal anyhow, and hence should *not* be guarded by the mutex.

Also, there's a pre-existing clang-tidy warning because of all the global variables which are illegal. We should make the registry a class and properly "static-initialize" it.

PiperOrigin-RevId: 214780540
GitOrigin-RevId: ae98b3f64affd4ca770fb31fa2c1c4d46086228c