Add a KeyManagerBase class in java, similar to the C++ class. Then, use it to implement the HmacKeyManager.
This gets rid of some of the boilerplate in each keymanager. However, we do add two more methods to parse key and keymanager in each instance. Nevertheless, it reduces much more code than it adds overall, and I think it's a clear simplification.

I marked the KeyManagerBase as alpha while it's in development.

PiperOrigin-RevId: 217838368
GitOrigin-RevId: fb8cbaa9ebf6663ecba5a862f197c952ded3758e