Add typesafe methods getKeyManager(), getPrimitive() and getPrimitives() to the registry.
In a follow up change, we will deprecate the old methods and migrate to them. Additionally, I use the new method in the AeadFactory to make sure it is used in a proguard test. In another follow up change, we will use it in all factories.

Finally, we add a method getUntypedKeyManager(), because it sometimes makes sense to have an untyped keymanager (when we just want to create the key from the template).

PiperOrigin-RevId: 218489845
GitOrigin-RevId: 350dc31d87be8a71a5ebac1edf78647765d0f454