Remove the semi-circular dependency between keyset_handle and registry.
Previously, the keyset_handle and registry would depend on each other, but in a way which could be handled. If we want to introduce the method keyset_handle.GetPrimitive<P>, doing this becomes at the very least very hard and confusing (I would have to split the header into two parts, essentially).

Instead, Bartosz pointed out that an option is to make the keyset_handle depend on the registry, and move the method GetPrimitives from the registry to the keyset_handle. I believe that this is the better option. However, it breaks clients who access GetPrimitives in the registry directly. There are no such clients within google; and I hope that there are non in the outside world either.

In any case, what needs to be done by any user is to replace all calls
  Registry::GetPrimitives(keyset_handle, keymanager);
with
  keyset_handle.GetPrimitives(keymanager);

PiperOrigin-RevId: 219093515
GitOrigin-RevId: efcd8ce66a94459a073b30f9818008a85904dc52