Java Tink: Add setwrappers to the registry.
This is the java analogue of the recent c++ change which added SetWrappers to the registry there.

This also changes the interface of PrimitiveSet<P> to have a method "Class<P> getPrimitiveClass();", this is needed because in the registry we want to look up the wrapper given a primitiveSet (and natural, anyhow).

PiperOrigin-RevId: 219110381
GitOrigin-RevId: 1004d26d3e3f20e93fbdc4bc6b8ee6410a3aabbb