Make the catalogue functions untyped.
The way the catalogue is used in the registry, there's actually not really a reason to use generics in this case. Hence we simply modify the functions to return the types they already return. Then we don't need to add a typeObject parameter.

The only time this will fail is if someone does something like

Catalogue<Aead> aeadCatalogue = Registry.getCatalogue("aead");

which, I'm pretty sure, nobody does.

PiperOrigin-RevId: 219467669
GitOrigin-RevId: dd17ac09b2df74eeaab3bf4e450305f56b726e9f