Update the javalite proto archive.
The javalite archive was updated on December 19: https://github.com/protocolbuffers/protobuf/commits/javalite

This means that the branch given by javalite.zip updates, so the hash changes. In principle, I would like to be able to check on my corp machine that this changes are exactly what I now have in the new machine. Unfortunately, I don't know how to download the previous archive with the sha256 hash to create a diff.

In order to make sure this works in the future, I suggest we download the archive by commit. For this, if you continue from the above link to the topmost commit, you get that the id of the link is 384...16 -- we copy this as the url and then we can get the corresponding sha256 sum on the local machine. If the branch updates, we need to manually update it.

The disadvantage is that we don't get cherry picks automatically.

This fixes issue 166.

PiperOrigin-RevId: 227678791
GitOrigin-RevId: cceb31b558c4fa9ca7613287cd69a8bad5f60798
(cherry picked from commit cd94de64cea4f7d7c5109fc84bd719940336644b)