Make keyset_handle->GetPrimitives() private.
Given that now keyset_handle has a function GetPrimitive, there seems to be no reason for it to expose GetPrimitives() anymore. Nobody should have called it, since the PrimitiveSet<> is not in the public APIs anyhow (and unless I'm missing something there's nothing you can give the result to immediately either or with auto).

The only exception would be if someone implemented their own factory, but I don't think there are factories at the moment.

PiperOrigin-RevId: 234436305
GitOrigin-RevId: f51581b2812c5ee03c7087acb63274eb6a2dc785