Implement a KeyManager based on an InternalKeyManager.
The main part of this cl is KeyManagerImpl.java, which implements the interface [1] based on the abstract class [2]. To construct the object, one then needs to provide a type parameter, specifying the generic parameter P in KeyManager<P> (the primitive).

We branch the test from the implementation of KeyManagerBase, since it's another implementation of the same interface, so we need to roughly test the same functions.

We need this cl since the registry exposes getKeyManager(), so we need to support the previous KeyManagers for a while.

PiperOrigin-RevId: 249240495