Add a function "firstSupportedPrimitiveClass" to the InternalKeyManager, and add support for InternalKeyManagers which do not have any primitive.
The function "firstSupportedPrimitiveClass" returns the class object of the first primitive which was configured, or java.lang.Void if no primitive was configured.

This is useful in the registry if the user asks for an UntypedKeyManager and we registered an InternalKeyMananger: we can return the key manager for the first type, and it's likely the same the user asked for previously.

PiperOrigin-RevId: 251214368