Replace std::tuple with a custom made template List in the InternalKeyManager.
std::tuple is a rather heavyweight type; for example it cannot be default-constructed. This created many hard to understand errors when I tried to add InternalKeyManager to the registry, some as weird that I would not be suprised if they were either standard library or compiler bugs. We can avoid all this by having our own type, which makes much more sense here anyhow.
PiperOrigin-RevId: 251617171