Add a SafeVararg warning suppression to the InternalKeyManager.
More explanation why the warning can be suppressed here can be found in Effective Java, Edition 3, Item 32. Short explanation:

When using variable arguments (...) in a method with a generic type, the compiler will create an array, containing the variable arguments. In our case, the array is of type PrimitiveFactory[] (without any generic type, the generic isn't part of this array).

Hence, the compiler would not complain if we added a PrimitiveFactory<?, SomethingWrong> to this array; in which case there would be all kinds of problems.

Since we don't do that the method is safe. As Effective Java puts it, it suffices to make sure that:
1) The constructor doesn't store anything in the vararg array parameter,
2) The constructor doesn't make the array visible to anyone.

Both conditions are satisfied here.

PiperOrigin-RevId: 252029999