Remove the Register(const KeyTypeEntry&) code and do what we do previously. Also, simply register everything in the Register(const RegistryConfig&) code.
The only way this can break someone we can think of is when the want to register fewer key managers, and have an implementation of one which they want to use instead, so a call to RegisterKeyManager now fails; this seems very unlikely and would be easy to detect and fix.

This should allow me to get rid of the config generation code.

PiperOrigin-RevId: 260902291