Add using declarations for the type parameters in KeyTypeManagers and use them.
We previously used passed an owned raw pointer from registry.h to registry_impl.h, because this makes it possible to infer template parameters (which doesn't work with unique pointers). However, it is better to instead specify the parameters explicitly, which is easiest if we add typedefs to the KeyTypeManager class.

This will also be useful elsewhere (to create [1] automatically).

PiperOrigin-RevId: 262295924