Add ABSL_MUST_USE_RESULT to Tink StatusOr.
Like in the corresponding commit pr/c6394fffd for Tink Status, We guard it with a "#ifndef CPP_TINK_TEMPORARY_STATUS_MUST_NOT_USE_RESULT", so that users can disable this if necessary.

Note that this also only can break the build for users which compile with both -Werror and -Wunused-result -- so another option for users to disable this is to remove -Wunused-result temporarily.

PiperOrigin-RevId: 263311358