Migrate from legacy 'java' provider to the JavaInfo provider.
'java' provider is deprecated and is going to be removed from bazel soon.

Also simplifying and fixing the code. Note that the "if hasattr(td, "java"):" branch was incorrect and was never accessed previously. Starlark didn't complain about it because one can call hasattr on any object. When trying to use the new provider it ended up in build errors (see rollback change cl/266444465) because the "in" operator is more restrictive and caught the bug.

PiperOrigin-RevId: 266745143