Manage Java deps with rules_jvm_external [1], and update rules_closure.
There are many ways to manage Java deps in Bazel [2], but rules_jvm_external is the only option that is officially maintained by the Bazel team. What we were using is unsupported and complicated. It also leads to dependency conflicts as described in https://github.com/google/tink/issues/189.

The new version of rules_closure and the migration to rules_jvm_external together resolve https://github.com/google/tink/issues/189, and make it possible to build a binary for Tink Javascript.

[1] https://github.com/bazelbuild/rules_jvm_external
[2] https://blog.bazel.build/2019/03/31/rules-jvm-external-maven.html

PiperOrigin-RevId: 267552697