Migrate the AesCtrHmacAeadKeyManager to a KeyTypeManager.
Note that this removes the call
  Registry.registerKeyManager(new AesCtrKeyManager(), /*newKeyAllowed=*/ true);
from the constructor. This is because I don't want the constructor to
be able to throw (else, accessing getKeyType() can be annoying).

Instead, we now access the HMacKeyManager and the AesCtrKeyManager directly (and not via the registry). Note that I need to make the HmacKeyManager public for that.

PiperOrigin-RevId: 267588791