Fix targets to avoid unstable string comparisons
The Message.String method is intended for debugging and does not guarantee
output stability. Code that relies on the output being stable for all time is
brittle and has unintentionally become a change detector for the proto package.
The v2 implementation of prototext (which will become the underlying
implementation of Message.String) will have deliberately unstable output
to prevent future code from relying on this unspecified property.

PiperOrigin-RevId: 273411890