Do not iterate through some of the depsets.
New bazel versions require us to get rid of code which is iterating through depsets; here we do some of this work for our proto compiler for objc.

Also, we pass in the compiler as a tool, instead of as an input. This is another incompatible bazel update (incompatible_no_support_tools_in_action_inputs)

PiperOrigin-RevId: 274978487