Upgrade to bazel 0.28.1.
We disable "bazel fetch" because it seems to do more than what is needed for build and test. Hence I'd rather only do the build.

We also update bazel-gazelle, as per  https://github.com/bazelbuild/bazel-gazelle#running-gazelle-with-bazel -- I'm not quite sure this is needed with disabling "bazel fetch", but it's a newer version so we should update anyhow.

PiperOrigin-RevId: 275047119