Move the classes TestUtil, StreamingTestUtil, and WycheproofTestUtil into the package com.google.crypto.tink.testing.
According to the best practices, only *Test.java should live under the testing tree. Because these classes are also in the way of me automating generation of BUILD files, I would like to move them to the non-testing tree. However, if we do that they will pollute the main Tink package, so we first put them into the "testing" subdirectory, and then later move them to the non-testing tree with the "testing" subdirectory.

PiperOrigin-RevId: 278833454