Add the guarantee to the Registry that KeyManagers returned by get_key_manager() remain valid, even in the presence of concurrent registering calls.
This was previously already the case, except for calls to RegisterAsymmetricKeyManagers after a non-asymmetric register call. Hence, there's a bit of a change of behavior, but if someone does this already I think they will prefer the new failure to the previous potential one.

I also remove the reference_wrapper TODO, since I think it's infeasible to change this at this point (see e.g. [1]). Let me know if you want to keep it.

PiperOrigin-RevId: 279289623