Add a method get_all to the C++ primitive_set class.
This is analogous to the java method getAll [1]. However, the return type is different since there's nothing as "Collection<List<Entry<P>>>" available anyhow. We could use a vector<vector<Entry<P>*>>, but I don't think that's an improvement.

I suggest simply to return a vector of entries in C++ instead.

PiperOrigin-RevId: 280157944