Add a java binary to compare keysets.
Because keysets can be serialized in different languages, I'd rather not compare the serialized version when comparing keysets (for example, go might decide that the fields are written in order in which they are declared in the proto file, while in Java they might be written in ascending order). I hence suggest to add such a binary.

(If some key data in some key manager contains another KeyData this might still fail, but I think we should ignore this for now).

NOTE: I specifically ignore the order of keys in the keyset; if you feel this is wrong, we should discuss it.
PiperOrigin-RevId: 283056426
(cherry picked from commit 859ed2c7a911968eef30e11777f0999de7f9384a)