Fix empty-x-and-y checking in validatePublicKey
Previously, this check incorrectly passed if x or y was an empty Uint8Array (a bytes proto field in JavaScript can be either a string or a Uint8Array at any given time).

The unit test that verifies this is tweaked from a null-checking test that was about to be deleted in an upcoming change anyway.

PiperOrigin-RevId: 284839725