Remove tests of runtime type checking from JavaScript version
These tests are an obstacle to migrating to TypeScript, as the type checker will flag them as type errors, and TypeScript disallows entirely disabling type checking as is currently done on these targets.

To speak briefly of the bigger picture: The current runtime type checking behavior is redundant if you assume that clients are using JSCompiler or TypeScript and not working around the type system, and insufficient if you don't assume that (they only check for a couple specific cases like null). Since we don't want to make that assumption, but also don't want to reduce the usefulness of the type checker for clients who do use it, in the medium term we should migrate to some kind of solution that automatically and rigorously does runtime type checking for all public APIs. This will hopefully obviate the need for each public API to have individual checks and test cases for those checks.

For now, the limited runtime type checks will remain in the implementation in order to avoid weakening the security of Tink. They won't have test cases, which is mildly unfortunate, but this is the easiest way to unblock the TypeScript migration and will hopefully be temporary.

Where it makes sense, the removed tests are replaced with ones that use an empty value of the correct type as a test input.

PiperOrigin-RevId: 284856694