BEGIN_PUBLIC Encrypt only DEK bytes for envelope encryption.
This changes the envelope encryption in C++ to solely encrypt the key
bytes of DEK instead of the full KeyData proto. This reduces the
overhead in the envelope encryption and fixes an issue with incompatible
format between C++ and the envelope encryption in Java and Go.

Cross languages tests for AEAD include now envelope encryption from C++.
END_PUBLIC

BUG=148046746

Change-Id: If502b497c7d3318f7163fbb20c2d714b1c1a4760