Add a StreamingPrfWrapper which wraps a single PRF into a single PRF.
This means that people who have a single raw key in the keyset can call KeysetHandle->GetPrimitive<StreamingPrf>() to create a Prf.

I think this is fine for the moment, and if we later create an interface which works for multiple keys people will be able to use that as well (in other words, this basically uses up the meaning of keyset_handle->GetPrimitive<StreamingPrf>(), but I think this provides the only meaning such a call can have.

PiperOrigin-RevId: 292302505