Encrypt only DEK bytes for envelope encryption.
This changes the envelope encryption in C++ to solely encrypt the key
bytes of the DEK instead of the full KeyData proto. This reduces the
overhead in the envelope encryption and fixes an issue with incompatible
format between C++ and the envelope encryption in Java and Go.

PiperOrigin-RevId: 292406414
(cherry picked from commit 02bdc3e37818d2367ee75e7835f0eb8129f857d5)