Inherit from native Error instead of goog.debug.Error
We're now using the techniques specified in https://javascript.info/custom-errors.

Most of the changed lines are in tests that check for particular error messages. This is a very non-robust pattern that we should really do something about at some point, but that can wait for a future change.

This is part of the effort to get rid of Closure Library dependencies.

PiperOrigin-RevId: 295875087