Roll forward of fba40f2bbbcfb3fe212c387c115dccb5d310d9c9: Inherit from native Error instead of goog.debug.Error
We're now using the techniques specified in https://javascript.info/custom-errors.

Most of the changed lines are in tests that check for particular error messages. This is a very non-robust pattern that we should really do something about at some point, but that can wait for a future change.

This is part of the effort to get rid of Closure Library dependencies.

This change was previously rolled back in 25b696e0a52e3278acc6060379d070f992517b6b.

NEW: Fix Google-internal clients that had been broken by this change as a result of depending on error messages. Also resolve many, many merge conflicts.
PiperOrigin-RevId: 297955343