Create new TypeScript public API
Currently, the API is just wrappers for underlying code implemented in Closure; follow-up changes will port it to TypeScript all the way down.

This is also the first preview of the new directory layout, which will be used for the underlying code following the TypeScript migration. If the layout works well, it may be used for other languages in the future.

PiperOrigin-RevId: 302060962