Change Tink for JavaScript primitive types from interfaces to abstract classes
TypeScript interfaces don't exist at runtime. This is a problem because APIs like KeysetHandle.getPrimitive take the primitive types as runtime arguments. We can solve this by using abstract classes instead of interfaces. This precludes inheriting from anything else, but that's not expected to be a problem in practice.

As an added bonus, this will facilitate more precise type checking of methods like getPrimitive, though we'll wait for the TypeScript migration to actually implement that.

PiperOrigin-RevId: 302433198