Migrate primitives from JavaScript to TypeScript
This translation was completed mostly by automated means (with the TypeScript output in particular coming from the gents tool), but some manual changes were needed to keep things building and the tests passing.

Additionally, the files providing primitives are being moved to internal-only directories, in order to keep each primitive out of builds that don't depend on it while avoiding circular Bazel dependencies. The primitive types are already reexported in the public API.

PiperOrigin-RevId: 304790700