Remove static async methods from Tink for JavaScript
These are blocking the TypeScript migration; due to a bug, tsickle transforms static async methods into code that the Closure Compiler doesn't know how to handle. Even if that bug were fixed, they still might not optimize as well as freestanding functions. In many cases, they also arguably aren't idiomatic; a lot of this code was originally a fairly direct port of Tink for Java, which doesn't have freestanding functions, but in JavaScript that's not an issue.

The APIs have names that are designed to make sense as qualified imports. Right now, from the caller's perspective, they still look a lot like static methods, because they're in files that default export a class. But that will no longer be true when those files are migrated to TypeScript, because then those classes will become named exports, and the new functions will be exported separately.

PiperOrigin-RevId: 312123753