Add option for defining base Tink workspace used for pip install.
Currently, pip package builds always depend on the latest version of the Tink master branch for building the C++ code with Bazel. This adds an option to set the Tink base directory used during the Bazel build of the Python extensions used for creating the pip package.

Updates the Tink package build test to not depend on the Github version anymore.

PiperOrigin-RevId: 312475411