Replace @TargetApi with @RequiresApi.
The TargetApi annotation is intended when there are errors in lint that make it believe that an API is only callable on a higher API version than it actually is. (That is, lint thinks the API is 23+ but it is actually 21+, for example)

The correct annotation to use is "RequiresApi" because the classes require the caller to be on API 23+ to make use of them.

PiperOrigin-RevId: 315356289